Simulation Early
The simulation works, but one thing it needs improvement on is the movement. If there is no available space to move into one direction - it doesn't check any other directions. Makes for an interesting view. But still works, although a winner will never be found as a numbers advantage can't be used by the infiltrated colony.